navigation and pages
there exists 3 main routes: main, faq, and join. main includes
subsections for content. currently these are just placeholders (as
are the other pages). we might change join to link to an external
service, for now, it is it's only template.

$anchorScroll was overriden so that the smooth scroll can be used
when hash changes occur within the main page. added jquery to
bower dependency.